refactor(auto_source): relocate Discovery and LinkHeuristics out of Scraper namespace
